grasping-Submission/external/r3m/requirements.txt
_libgcc_mutex=0.1
_openmp_mutex=4.5
bzip2=1.0.8
ca-certificates=2021.10.8
ld_impl_linux-64=2.36.1
libffi=3.4.2
libgcc-ng=11.2.0
libgomp=11.2.0
ibnsl=2.0.0
libuuid=2.32.1
libzlib=1.2.11
ncurses=6.3
openssl=3.0.0
pip=22.0.4
python=3.9.10
python_abi=3.9
readline=8.1
setuptools=60.9.3
sqlite=3.37.0
tk=8.6.12
tzdata=2021e
wheel=0.37.1
xz=5.2.5
lib=1.2.11
antlr4-python3-runtime==4.8
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
filelock==3.6.0
fonttools==4.30.0
gdown==4.4.0
huggingface-hub==0.4.0
hydra-core==1.1.1
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.3
omegaconf==2.1.1
packaging==21.3
pillow==9.0.1
pyparsing==3.0.7
pysocks==1.7.1
python-dateutil==2.8.2
pyyaml==6.0
regex==2022.3.2
requests==2.27.1
sacremoses==0.0.47
six==1.16.0
soupsieve==2.3.1
tokenizers==0.11.6
torch==1.7.1
torchvision==0.8.2
tqdm==4.63.0
transformers==4.17.0
typing-extensions==4.1.1
urllib3==1.26.8
